Add prometheus instrumentation library
This is a 3rd party hook implementation which adds a hook that records
metrics for all registered twirp handlers automatically.

We also add the http handler to our mux so that the metrics are exposed
such that they can be collected by a prometheus collector.

This might not be suitable for real deployment, but adding for now so
that we can see it operating from the outside.