Remove retryable and initialize on new server
Previously we created the registerer in an init call in the global scope
but that caused problems in the event we had to retry initialization while
waiting for the db server to come online. This change attempts to
initialize a new registry inside the NewServer method and not in init so
that it should be empty on each retry.